chore: 도메인 로직을 domain 폴더 내부로 변경
