+++ src/domain/CarManager.js
@@ -39,18 +39,6 @@ class CarManager {
       console.log(); // 줄바꿈
     }
   }
-
-  determineWinners() {
-    const carsPosition = this.cars.map((car) => car.position);
-
-    const maxPosition = Math.max(...carsPosition);
-
-    const winners = this.cars
-      .filter((car) => car.position === maxPosition)
-      .map((car) => car.name);
-
-    return winners;
-  }
 }
 
 export default CarManager;
